get scm credential
