+++ awx_collection/plugins/module_utils/export_tools.py
@@ -14,3 +14,16 @@ def is_object_in_list_by_id(object_id, list):
         if object['id'] == object_id:
             return True
     return False
+
+def transform_users_set_to_objects(users_info_set):
+    users = []
+    for user_info in users_info_set:
+        user_array = user_info.split(';')
+        user = {
+            'username': user_array[0],
+            'first_name': user_array[1],
+            'last_name': user_array[2],
+            'email': user_array[3]
+        }
+        users.append(user)
+    return users
